Rename harp-cell-wrapper as harp-cell
There's two benefits to this:

1. Wrapper is not a particularly informative name if all it's denoting
is a hierarchy. The hierachy is implied strongly enough under this
convention
2. The pull request diff will be much more clear in it's highlights of
the *alteractions* introduced in this change. Functionality has been
drawn out from harp-cell *more* than harp-cell has been replaced. This
will make that clearer.
