--- packages/harpguru-core/src/components/harp-cell-wrapper/index.ts
@@ -1,2 +1,0 @@
-export { HarpCellWrapper } from './harp-cell-wrapper'
-export type { YXCoord } from './harp-cell-wrapper'
