+++ packages/harpguru-core/src/components/harp-cell/hooks/use-position-analysis/index.ts
