+++ packages/harpguru-core/src/components/harp-cell/hooks/use-position-analysis/use-position-analysis.test.ts
@@ -1,7 +1,7 @@
 import { useGlobal } from 'reactn'
 import type { Degree, IsActiveIds, Pitch } from 'harpstrata'
 
-import type { YXCoord } from '../../harp-cell-wrapper'
+import type { YXCoord } from '../../harp-cell'
 import { inactiveCellsHarpStrata as activeHarpStrata } from '../../../../test-resources'
 
 import { usePositionAnalysis } from './index'
