+++ packages/harpguru-core/src/components/harp-cell/index.ts
@@ -1,0 +1,2 @@
+export { HarpCell } from './harp-cell'
+export type { YXCoord } from './harp-cell'
